refactor(shared components): Rename 'fxa-components' to 'fxa-react'
Because:
We need to begin sharing at least hooks and images alongside components. Components should live in fxa-react/components/ instead of root-level fxa-components.
