+++ packages/fxa-admin-panel/.rescriptsrc.js
@@ -3,7 +3,7 @@ const {
   permitAdditionalJSImports,
   setupAliasedPaths,
   componentsJestMapper,
-} = require('../fxa-components/rescripts');
+} = require('../fxa-react/rescripts');
 
 module.exports = [
   permitAdditionalJSImports,
