+++ packages/fxa-admin-panel/package.json
@@ -74,7 +74,7 @@
     "eslint-plugin-fxa": "^2.0.1",
     "eslint-plugin-jest": "^23.8.1",
     "eslint-plugin-react": "^7.18.3",
-    "fxa-components": "file:../fxa-components",
+    "fxa-react": "file:../fxa-react",
     "pm2": "^4.2.3",
     "prettier": "^1.19.1",
     "supertest": "^4.0.2",
