+++ packages/fxa-admin-panel/tsconfig.json
@@ -31,5 +31,5 @@
   "include": [
     "src"
   ],
-  "extends": "../fxa-components/tsconfig.common.json"
+  "extends": "../fxa-react/tsconfig.common.json"
 }
