+++ packages/fxa-react/.eslintrc
