+++ packages/fxa-react/.storybook/main.js
