+++ packages/fxa-react/.storybook/webpack.config.js
