+++ packages/fxa-react/LICENSE
