+++ packages/fxa-react/README.md
