+++ packages/fxa-react/components/AppErrorBoundary/index.test.tsx
