+++ packages/fxa-react/components/AppErrorBoundary/index.tsx
