+++ packages/fxa-react/components/AppErrorDialog/index.stories.tsx
