+++ packages/fxa-react/components/AppErrorDialog/index.test.tsx
