+++ packages/fxa-react/components/AppErrorDialog/index.tsx
