+++ packages/fxa-react/components/LinkExternal/index.test.tsx
