+++ packages/fxa-react/components/LinkExternal/index.tsx
