+++ packages/fxa-react/components/Survey/index.scss
