+++ packages/fxa-react/components/Survey/index.stories.tsx
