+++ packages/fxa-react/components/Survey/index.test.tsx
