+++ packages/fxa-react/components/Survey/index.tsx
