+++ packages/fxa-react/jest.config.js
