+++ packages/fxa-react/tsconfig.json
