+++ packages/fxa-settings/.rescriptsrc.js
@@ -6,7 +6,7 @@ const {
   permitAdditionalJSImports,
   setupAliasedPaths,
   componentsJestMapper,
-} = require('../fxa-components/rescripts');
+} = require('../fxa-react/rescripts');
 
 module.exports = [
   permitAdditionalJSImports,
