+++ packages/fxa-settings/tsconfig.json
@@ -22,5 +22,5 @@
   "include": [
     "src"
   ],
-  "extends": "../fxa-components/tsconfig.common.json"
+  "extends": "../fxa-react/tsconfig.common.json"
 }
